refactor(signatures): move error types from minisign.rs to keys.rs
Removes dependency of the error types on minisign.
Better design, needed to support signing alternatives, and for
https://github.com/asfaload/asfasign/pull/13#discussion_r2311949399
